% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLFRED.R
\name{getFRED}
\alias{getFRED}
\title{Get a time-series from the FRED}
\usage{
getFRED(series_id)
}
\arguments{
\item{series_id}{The series id for the time-series to be downloaded}
}
\description{
Downloads a time-series from FRED into R
}
\details{
The data is downloaded from the FRED API
}
\author{
Graeme Walsh
}
