% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLFRED.R
\name{getObs}
\alias{getObs}
\title{Get time-series values (observations) from the FRED}
\usage{
getObs(series_id)
}
\arguments{
\item{series_id}{The series id for the time-series to be downloaded}
}
\description{
Downloads time-series values (observations) from FRED into R
}
\details{
The data is downloaded from the FRED API
}
\author{
Graeme Walsh
}
