% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLFRED.R
\name{getSeriesMeta}
\alias{getSeriesMeta}
\title{Get time-series metadata from the FRED}
\usage{
getSeriesMeta(series_id)
}
\arguments{
\item{series_id}{The series id for the time-series of interest}
}
\description{
Downloads time-series metadata from FRED into R
}
\details{
The data is downloaded from the FRED API
}
\author{
Graeme Walsh
}
