% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLFRED.R
\name{getSLFRED2}
\alias{getSLFRED2}
\title{Get a time-series from the FRED}
\usage{
getSLFRED2(series_id)
}
\arguments{
\item{series_id}{The series id for the time-series to be downloaded}
}
\description{
Downloads a time-series from FRED into R
}
\details{
The data is downloaded from the FRED data folder (not using the API).
}
\author{
Graeme Walsh
}
