% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLFRED.R
\docType{package}
\name{SLFRED}
\alias{SLFRED}
\alias{SLFRED-package}
\title{A package for downloading time-series from the FRED database.}
\description{
The workhorse function of the package is \code{getFRED()}. This function returns a
list containing the requested time-series and the corresponding metadata.
The code behind this package also serves as a backend to an EViews add-in created by the author.
}
\section{SLFRED functions}{

\code{getFRED()}
}

